ENH: AttenuatorBase can now be used for the lightpath
For this implementation I wrote methods that would work on either the FEE or
standard LusiAttenuator IOCs. This means things are done slightly inefficently
for the newer IOC as more records exist to indicate the state. However, I
think in the interest of time this is the right choice for now. We can go back
whenever and either reimplement some of the methods for the newer version, or
replace the IOC in the FEE.