ENH: Create basic GateValve
The GateValve and Stopper were so similar that I just had the GateValve be a
reimplementation of the Stopper class. This reduced the amount of duplicate
code by a great deal. Someday we may want ot just make a `LimitSwitch` state
object and have Stopper and GateValve both be separate subclasses, but for now
this will work