FIX: Create PointingMirror class
In order to accomodate mirrors that do not point between beamlines, these have
now been split in between two classes; OffsetMirror and PointingMirror. The
former does not insert and retract, instead it is always assumed this mirror is
in the beampath. PointingMirror has subcomponents for the MPS system as well as
the state record for mirrors that need to inserted for alignment.