API: Remove mps logic from devices where MPS is not mandatory
I removed MPS logic from GateValve and Stopper and created the classes
MPSGateValve and MPSStopper using the factory function. I left the logic in
both PointingMirror and PPSStopper as these devices should always have an MPS
bit.  I am also breaking API for the sake of consistency. The mps prefix is
always specified with the keyword `mps_prefix`, regardless whether it is baked
into the class or built using the factory

so we do it consistenly across all devices with