CI: Fix issue where package was not uploaded
In the test environment, we didn't have anaconda-client installed.
This meant that at the end of the tests we were unable to upload
to anaconda. This commit makes sure we're in the correct conda env
and defers the install of anaconda-client until the very end
since it's only needed if we are uploading.