Added an explicit ``set_many()`` check that the exception is related.
In ``set_many()`` we catch a TypeError to see if the *settle_time*
parameter is accepted by each signal's ``set()`` method, and if it is
we try again without the *settle_time* argument. If a TypeError is
raised for another reason, this would also retry the ``set()``
call. We can't guarantee that this is indempotent. Now we explicity
check that the TypeError is at least related to the *settle_time*
argument before retrying.